# doc/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  version: 0.0.5
  title: Notch Health D2C 
  description: >
    ## Overview

    This specification defines the necessary resources needed to access a patient's information and the patient's test results.
    
    
    A customer orders a test kit through a website, when the test kit order is received by the customer, they register each kit to a person who will take the test. When registering the kit, they will enter a test kit ID along with their name, date of birth and other personal information. 
    
    Upon submission of the kit ID, a lab order is created. Once the customer has administered the test, they mail back the kit to the lab.
    
    
    The customer can track the kit to the lab, check the status or the results, update any personal information related to the lab order.
    
    
    Eight services exist to that act as a broker of information between the lab system and the website.
    
    1) A service to create a kit order request
    
    2) A service to request the delivery status of a kit
    
    3) A service to update the kit order with kit IDs
    
    4) A service to create a lab order
    
    5) A service to get all lab orders for a customer
    
    6) A service to get a specific lab order (to check the status)
    
    7) A service to update a lab order
    
    8) A service to delete a lab order
    
    9) A service to check the existence of a lab order

    10) A service to get the customer's Shopify orders

...

...

...

...

...


...


...


...


...

...

...
host: "notchapi-service-ovvelzitca-uw.a.run.app"
basePath: / 
schemes:
...
produces:
...
consumes:
...
paths:
...
    /lab-orders:
      post:
        tags:
          - lab-order
        parameters:
          - name: lab-order
            in: body
            description: "User events captured"
            required: true
            schema:
              type: object
              items:
                $ref: '#/definitions/lab-order'
        summary: "Create a lab order for the customer"
        operationId: labOrderPost
        responses:
          '201':
            description: "lab order object created "
      get:
        tags:
          - lab-order
        parameters:
          - name: customer-id
            in: query
            description: "User events captured"
            required: true
            type: string
        summary: "Get all lab orders for customer based on customer ID"
        operationId: labOrderGet
        responses:
          '200':
            description: "lab order response object"
            schema:
              type: object
              items:
                $ref: '#/definitions/lab-order'
    /lab-orders/{id}:
      get:
        tags:
          - lab-order
        summary: "Get a specified lab order based on order ID"
        operationId: labOrderGetByID
        parameters:
          - name: id
            in: path
            description: "ID of the lab order"
            required: true
            type: string
        responses:
          '200':
            description: "delivery response object"
      put:
        tags:
          - lab-order
        parameters:
          - name: id
            in: path
            description: "ID of the lab order"
            required: true
            type: string
          - name: lab-order
            in: body
            description: "User events captured"
            required: true
            schema:
              type: object
              items:
                $ref: '#/definitions/lab-order'
        summary: "Update the lab order for a customer/practioner based on order ID"
        operationId: labOrderUpdateFullByID
        responses:
          '200':
            description: "updated lab order"
      delete:
        tags:
          - lab-order
        parameters:
          - name: id
            in: path
            description: "ID of the lab order"
            required: true
            type: string
        summary: "Delete the lab order based on order ID"
        operationId: labOrderDelete
        responses:
          '200':
            description: "deletion successful"
      head:
        tags:
          - lab-order
        parameters:
          - name: id
            in: path
            description: "ID of the lab order"
            required: true
            type: string
        summary: "See if lab order exist based on order ID"
        operationId: labOrderHead
        responses:
          '200':
            description: "Lab order exist"
    /orders/{id}:
      get:
        tags:
          - order
        parameters:
          - name: id
            in: path
            description: "ID of the shopify order"
            required: true
            type: string
        summary: "Get a specified shopify order based on order ID"
        operationId: ShopifyOrderGetByID
        responses:
          '200':
            description: "response object"
    /orders:
      post:
        tags:
          - order
        parameters:
          - name: email
            in: body
            description: "email address associated to the shopify order"
            required: true
            schema:
              type: string
        summary: "Get a specified shopify order based on order email"
        operationId: ShopifyOrderGetByEmail
        responses:
          '200':
            description: "response object"
    /customers/{id}:
      get:
        tags:
          - customer
        parameters:
          - name: id
            in: path
            description: "ID of the shopify order"
            required: true
            type: string
        summary: "Get a specified shopify customer based on customer ID"
        operationId: ShopifyCustomerGetByID
        responses:
          '200':
            description: "response object"
    /customers:
      post:
        tags:
          - customer
        parameters:
          - name: email
            in: body
            description: "Email address of the shopify customer"
            required: true
            schema:
              type: string
        summary: "Get a specified shopify customer based on customer ID"
        operationId: ShopifyCustomerGetByEmail
        responses:
          '200':
            description: "response object"
    /shopfiy-orders/{id}:
      get:
        tags:
          - shopify-direct
        parameters:
          - name: id
            in: path
            description: "ID address of shopify order"
            required: true
            type: string
        summary: "Get a specified shopify order from Shopify"
        operationId: GetOrderFromShopiy
        responses:
          '200':
            description: "response object"
...
definitions:
...
  lab-order:
    type: object
    properties:
      id:
        type: string
      order-id:
        type: string
      customer-id:
        type: string
      prefix:
        type: string
      invoice-num:
        type: string
      invoice-date:
        type: string
      prac-id:
        type: string
      prac-name:
        type: string
      print-batch-id:
        type: string
      loc-pos:
        type: string
      owner-fname:
        type: string
      owner-mname:
        type: string
      owner-lname:
        type: string
      patient-fname:
        type: string
      patient-mname:
        type: string
      patient-lname:
        type: string
      patient-gender:
        type: string
      patient-age:
        type: string
      patient-dob:
        type: string
      patient-ssn:
        type: string
      patient-address:
        type: string
      patient-city:
        type: string
      patient-state:
        type: string
      patient-country:
        type: string
      patient-zip:
        type: string
      patient-phday:
        type: string
      patient-pheve:
        type: string
      ship-method:
        type: string
      tracking-num:
        type: string
      ship-priority:
        type: string
      date-received:
        type: string
      date-drawn:
        type: string
      species:
        type: string
      breed:
        type: string
      sample-type:
        type: string
      sample-volume:
        type: string
      patient-phase:
        type: string
      patient-cycle-day:
        type: string
      patient-cycle-length:
        type: string
      storage:
        type: string
      diagnoses:
        type: string
      icd9:
        type: string
      test-ordered:
        type: string
      locale:
        type: string
      test-priority:
        type: string
      rush-date:
        type: string
      bill-method:
        type: string
      bill-ref:
        type: string
      check-amount:
        type: string
      status:
        type: string
      report-template:
        type: string
      comment:
        type: string
      entry-comment:
        type: string
      reject-code:
        type: string
      processing-comment:
        type: string
      ordered-date:
        type: string
      ordered-by:
        type: string
      request:
        type: string
      do-not-send:
        type: string
      verified:
        type: string
      sent-to-blis:
        type: string
      sent-to-qb:
        type: string
      label-amount:
        type: string
      label-print-status:
        type: string
      override-partner-processing:
        type: string
      partner-id:
        type: string
      partner-status-id:
        type: string
      partner-processing-comment:
        type: string
      time-drawn:
        type: string
      time-zone-drawn:
        type: string
      partner-order-id:
        type: string
      approved-order-type-id:
        type: string
      file:
        type: string